position event context at page level
